+++ test/mstl_test.cpp
@@ -16,19 +16,19 @@ void test_mstl_works() {
     MstlResult<T> result = stl::mstl_params().fit(generate_series<T>(), {6, 10});
     assert_elements_in_delta(
         {0.28318232, 0.70529824, -1.980384, 2.1643379, -2.3356874},
-        first(result.seasonal.at(0), 5)
+        first(result.seasonal().at(0), 5)
     );
     assert_elements_in_delta(
         {1.4130436, 1.6048906, 0.050958008, -1.8706754, -1.7704514},
-        first(result.seasonal.at(1), 5)
+        first(result.seasonal().at(1), 5)
     );
     assert_elements_in_delta(
         {5.139485, 5.223691, 5.3078976, 5.387292, 5.4666862},
-        first(result.trend, 5)
+        first(result.trend(), 5)
     );
     assert_elements_in_delta(
         {-1.835711, 1.4661198, -1.3784716, 3.319045, -1.3605475},
-        first(result.remainder, 5)
+        first(result.remainder(), 5)
     );
 }
 
@@ -38,19 +38,19 @@ void test_mstl_span() {
     MstlResult<T> result = stl::mstl_params().fit(std::span<const T>(series), {{6, 10}});
     assert_elements_in_delta(
         {0.28318232, 0.70529824, -1.980384, 2.1643379, -2.3356874},
-        first(result.seasonal.at(0), 5)
+        first(result.seasonal().at(0), 5)
     );
     assert_elements_in_delta(
         {1.4130436, 1.6048906, 0.050958008, -1.8706754, -1.7704514},
-        first(result.seasonal.at(1), 5)
+        first(result.seasonal().at(1), 5)
     );
     assert_elements_in_delta(
         {5.139485, 5.223691, 5.3078976, 5.387292, 5.4666862},
-        first(result.trend, 5)
+        first(result.trend(), 5)
     );
     assert_elements_in_delta(
         {-1.835711, 1.4661198, -1.3784716, 3.319045, -1.3605475},
-        first(result.remainder, 5)
+        first(result.remainder(), 5)
     );
 }
 
@@ -59,19 +59,19 @@ void test_mstl_unsorted_periods() {
     MstlResult<T> result = stl::mstl_params().fit(generate_series<T>(), {10, 6});
     assert_elements_in_delta(
         {1.4130436, 1.6048906, 0.050958008, -1.8706754, -1.7704514},
-        first(result.seasonal.at(0), 5)
+        first(result.seasonal().at(0), 5)
     );
     assert_elements_in_delta(
         {0.28318232, 0.70529824, -1.980384, 2.1643379, -2.3356874},
-        first(result.seasonal.at(1), 5)
+        first(result.seasonal().at(1), 5)
     );
     assert_elements_in_delta(
         {5.139485, 5.223691, 5.3078976, 5.387292, 5.4666862},
-        first(result.trend, 5)
+        first(result.trend(), 5)
     );
     assert_elements_in_delta(
         {-1.835711, 1.4661198, -1.3784716, 3.319045, -1.3605475},
-        first(result.remainder, 5)
+        first(result.remainder(), 5)
     );
 }
 
@@ -80,19 +80,19 @@ void test_mstl_lambda() {
     MstlResult<T> result = stl::mstl_params().lambda(0.5).fit(generate_series<T>(), {6, 10});
     assert_elements_in_delta(
         {0.43371448, 0.10503793, -0.7178911, 1.2356076, -1.8253292},
-        first(result.seasonal.at(0), 5)
+        first(result.seasonal().at(0), 5)
     );
     assert_elements_in_delta(
         {1.0437742, 0.8650516, 0.07303603, -1.428663, -1.1990008},
-        first(result.seasonal.at(1), 5)
+        first(result.seasonal().at(1), 5)
     );
     assert_elements_in_delta(
         {2.0748303, 2.1291165, 2.1834028, 2.2330272, 2.2826517},
-        first(result.trend, 5)
+        first(result.trend(), 5)
     );
     assert_elements_in_delta(
         {-1.0801829, 0.900794, -0.7101207, 1.9600279, -1.2583216},
-        first(result.remainder, 5)
+        first(result.remainder(), 5)
     );
 }
 
@@ -105,19 +105,19 @@ void test_mstl_lambda_zero() {
     MstlResult<T> result = stl::mstl_params().lambda(0.0).fit(series, {6, 10});
     assert_elements_in_delta(
         {0.18727916, 0.029921893, -0.2716494, 0.47748315, -0.7320051},
-        first(result.seasonal.at(0), 5)
+        first(result.seasonal().at(0), 5)
     );
     assert_elements_in_delta(
         {0.42725056, 0.32145387, -0.019030934, -0.56607914, -0.46765903},
-        first(result.seasonal.at(1), 5)
+        first(result.seasonal().at(1), 5)
     );
     assert_elements_in_delta(
         {1.592807, 1.6144379, 1.6360688, 1.6559447, 1.6758206},
-        first(result.trend, 5)
+        first(result.trend(), 5)
     );
     assert_elements_in_delta(
         {-0.41557717, 0.33677137, -0.24677622, 0.7352363, -0.47615635},
-        first(result.remainder, 5)
+        first(result.remainder(), 5)
     );
 }