+++ test/stl_test.cpp
@@ -17,17 +17,17 @@ void test_stl_works() {
     StlResult<T> result = stl::params().fit(series, 7);
     assert_elements_in_delta(
         {0.36926576, 0.75655484, -1.3324139, 1.9553658, -0.6044802},
-        first(result.seasonal, 5)
+        first(result.seasonal(), 5)
     );
     assert_elements_in_delta(
         {4.804099, 4.9097075, 5.015316, 5.16045, 5.305584},
-        first(result.trend, 5)
+        first(result.trend(), 5)
     );
     assert_elements_in_delta(
         {-0.17336464, 3.3337379, -1.6829021, 1.8841844, -4.7011037},
-        first(result.remainder, 5)
+        first(result.remainder(), 5)
     );
-    assert_elements_in_delta({1.0, 1.0, 1.0, 1.0, 1.0}, first(result.weights, 5));
+    assert_elements_in_delta({1.0, 1.0, 1.0, 1.0, 1.0}, first(result.weights(), 5));
 }
 
 template<typename T>
@@ -36,17 +36,17 @@ void test_stl_span() {
     StlResult<T> result = stl::params().fit(std::span<const T>(series), 7);
     assert_elements_in_delta(
         {0.36926576, 0.75655484, -1.3324139, 1.9553658, -0.6044802},
-        first(result.seasonal, 5)
+        first(result.seasonal(), 5)
     );
     assert_elements_in_delta(
         {4.804099, 4.9097075, 5.015316, 5.16045, 5.305584},
-        first(result.trend, 5)
+        first(result.trend(), 5)
     );
     assert_elements_in_delta(
         {-0.17336464, 3.3337379, -1.6829021, 1.8841844, -4.7011037},
-        first(result.remainder, 5)
+        first(result.remainder(), 5)
     );
-    assert_elements_in_delta({1.0, 1.0, 1.0, 1.0, 1.0}, first(result.weights, 5));
+    assert_elements_in_delta({1.0, 1.0, 1.0, 1.0, 1.0}, first(result.weights(), 5));
 }
 
 template<typename T>
@@ -55,19 +55,19 @@ void test_stl_robust() {
     StlResult<T> result = stl::params().robust(true).fit(series, 7);
     assert_elements_in_delta(
         {0.14922355, 0.47939026, -1.833231, 1.7411387, 0.8200711},
-        first(result.seasonal, 5)
+        first(result.seasonal(), 5)
     );
     assert_elements_in_delta(
         {5.397365, 5.4745436, 5.5517216, 5.6499176, 5.748114},
-        first(result.trend, 5)
+        first(result.trend(), 5)
     );
     assert_elements_in_delta(
         {-0.5465884, 3.0460663, -1.7184906, 1.6089439, -6.5681853},
-        first(result.remainder, 5)
+        first(result.remainder(), 5)
     );
     assert_elements_in_delta(
         {0.99374926, 0.8129377, 0.9385952, 0.9458036, 0.29742217},
-        first(result.weights, 5)
+        first(result.weights(), 5)
     );
 }