+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.3.0 (unreleased)
 
+- Changed `seasonal`, `trend`, `remainder`, and `weights` to functions
 - Dropped support for C++17
 
 ## 0.2.0 (2025-02-23)